desafio3/web/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1

